+++ CHANGES.md
@@ -1,5 +1,10 @@
 # CHANGES for csv-to-indexeddb
 
+## ?
+
+- Linting (ESLint): utilize chai-friendly/chai-expect recommended configs
+- npm: Update devDeps
+
 ## 0.0.1
 
 - Initial release